% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by.R
\name{count_by}
\alias{count_by}
\title{Count value by variables}
\usage{
count_by(data, grp, var, var_label = var, id = "USUBJID")
}
\arguments{
\item{data}{A data frame}

\item{grp}{Group variable in \code{data}.}

\item{var}{Analysis variable in \code{data}.}

\item{var_label}{A string of analysis variable label.}

\item{id}{Subject id variable in \code{data}.}
}
\description{
Count value by variables
}
